% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMcluster-package.R
\docType{package}
\name{MMcluster-package}
\alias{MMcluster}
\alias{MMcluster-package}
\title{MMcluster: Mixed-Model Clustering with a Gibbs Sampler}
\description{
Performs mixture model clustering using a Gibbs sampler by assuming the data follows a random effects mixture model.
}
\author{
\strong{Maintainer}: Gregory Reid \email{gj3reid@edu.uwaterloo.ca}

Authors:
\itemize{
  \item Jose Carlos Dairo \email{jcfdairo@edu.uwaterloo.ca}
  \item Thomas Yan \email{tsyan@edu.uwaterloo.ca}
  \item Mark Christopher Uy \email{mark.uy@edu.uwaterloo.ca}
}

}
\keyword{internal}
