% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMcluster-function.R
\name{MMcluster}
\alias{MMcluster}
\title{Mixture Model Clustering}
\usage{
MMcluster(
  data,
  V,
  prior,
  initParamVals,
  K,
  burnin_period,
  numIter,
  mu.fixed = FALSE,
  Sigma.fixed = FALSE,
  theta.fixed = FALSE,
  rho.fixed = FALSE,
  z.fixed = FALSE,
  Theta.out = FALSE,
  z.out = FALSE
)
}
\arguments{
\item{data}{Input matrix with dimensions \verb{N x p}.}

\item{V}{The observed Fisher information matrix of each observation. Given as array with dimensions \verb{p x p x N}.}

\item{prior}{An optional list containing the parameters for the prior selected. Must have elements named \code{Vk} and \code{Omega}.
\code{Vk} being a vector of length \code{K} and \code{Omega} being an array of variance matricies with dimension \verb{p x p x K}.}

\item{initParamVals}{A list containing the initial values of the parameters of the model. Must have elements names \code{theta}, \code{mu}, \code{Sigma}, \code{rho}, and \code{z}. Where
\code{theta} is a \verb{N x p} matrix,
\code{mu} is a \verb{K x p} matrix,
\code{Sigma} is a \verb{p x p x K} array of variance matricies,
\code{rho} is a probability vector with length \code{K},
and \code{z} is a vector of length \code{N} with each element being an integer within \{1,...,K\}.}

\item{K}{A scalar value indicating the desired number of clusters.}

\item{burnin_period}{A scalar value indicating the initial number of samples to discard.}

\item{numIter}{A scalar value indicating the desired number of MCMC samples to be generated.}

\item{mu.fixed}{A true or false indicating whether to sample new mu values. See details for the usage of the mu parameter.}

\item{Sigma.fixed}{A true or false indicating whether to sample new Sigma values. See details for the usage of the Sigma parameter.}

\item{theta.fixed}{A true or false indicating whether to sample new theta values. See details for the usage of the theta parameter.}

\item{rho.fixed}{A true or false indicating whether to sample new rho values. See details for the usage of the rho parameter.}

\item{z.fixed}{A true or false indicating whether to sample new z values. See details for the usage of the z parameter.}

\item{Theta.out}{A true or false indicating whether to save and return all Theta samples. Caution: This will be a very large object.}

\item{z.out}{A true or false indicating whether to save all z samples. Caution: This will be a very large object.}
}
\value{
A list of all the sampled parameters to the posterior distribution of z. The named elements are as follows: \cr
\code{theta} is a \verb{N x p} matrix or Theta.out = TRUE a \verb{N x p x M} array. \cr
\code{mu} is a \verb{K x p x M} array. \cr
\code{Sigma} is a \verb{p x p x K x M} array of variance matricies. \cr
\code{rho} is a \verb{K x M} matrix with the rows being probability vectors (sum to 1). \cr
\code{z} is a length \code{N} vector with each element being an integer within \{1,...,K\} or if z.out = TRUE a \verb{N x M} matrix.
}
\description{
Performs mixture model clustering using a Gibbs sampler by assuming the data follows a random effects mixture model.
}
\details{
The hierarchical model used in the Gibbs sampler is as follows:
\deqn{z ~ Multinom(K, \rho)}
Where K is the number of clusters and \eqn{\rho} is a probability vector for being in each group. Each z is identically independently distributed.
\deqn{\theta | z ~ N(\mu, \Sigma)}
Where there are K different \eqn{\mu} and \eqn{\Sigma} parameters, one each for every cluster. This cluster is given by the cluster observation of z.
\deqn{y | \theta ~ N(\theta, V)}
Where y is the data given to the function and V is the Fisher information matrix of each observation. So, \eqn{\theta} creates a random effects model of the data for y.
}
