% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_info.R
\docType{data}
\name{obs_info}
\alias{obs_info}
\title{Fisher Information for HBE Data}
\format{
A 6 x 6 x 668 array of covariance matricies
}
\source{
Hill, David B et al. “A biophysical basis for mucus solids concentration as a candidate biomarker for airways disease.” PloS one vol. 9,2 e87681. 18 Feb. 2014, doi:10.1371/journal.pone.0087681
}
\usage{
obs_info
}
\description{
A dataset including the Fisher information matricies for the HBE dataset.
}
\keyword{datasets}
