% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{solveV}
\alias{solveV}
\title{Solve method for variance matrices.}
\usage{
solveV(V, x, ldV = FALSE)
}
\arguments{
\item{V}{Variance matrix.}

\item{x}{Optional vector or matrix for which to solve system of equations.  If missing calculates inverse matrix.}

\item{ldV}{Optionally compute log determinant as well.}
}
\value{
Either a matrix or vector \verb{y = V^\{-1\} x}, or if \code{ldV = TRUE}, a list with elements \code{y} and \code{ldV = log(det(V))}.
}
\description{
Calculates \verb{y = V^\{-1\} x} where \code{V} is a variance matrix.
}
\details{
This function is faster and more stable than \code{base::solve()} when \code{V} is known to be positive-definite.
}
\author{
Martin Lysy
}
