% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_rho}
\alias{update_rho}
\title{Samples cluster memberships probabilities from a Dirichlet distribution}
\usage{
update_rho(theta, mu, Sigma, Z)
}
\arguments{
\item{theta}{An 'n x p' matrix, where row i is the parameters of observation i.}

\item{mu}{An 'K x p' matrix where the row k is expected value of theta if theta is in class k.}

\item{Sigma}{A 'p x p x K' array of variance-covariance matrices for each class k.}

\item{Z}{An 'n x 1' vector of class memberships for each observation}
}
\value{
An 'n x 1' vector of updated class membership probabilities.
}
\description{
Samples cluster memberships probabilities from a Dirichlet distribution
}
