% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_Sigma}
\alias{update_Sigma}
\title{Samples an array of matrices denoting the covariance matrices for each cluster}
\usage{
update_Sigma(y, theta, mu, old_Sigma, z, Omega, vK)
}
\arguments{
\item{y}{A \verb{n x p} matrix. Each row of y is an observation on p dimensions of the data.}

\item{theta}{A \verb{n x p} matrix where each row is an observation and the columns are a subset of a larger parameter set.}

\item{mu}{An \verb{K x p} vector denoting the prior cluster means.}

\item{old_Sigma}{A \verb{p x p x K} array of variance matricies. The previous observation of Sigma.}

\item{z}{A length \code{n} vector denoting the class that each row of theta belongs to.}

\item{Omega}{The covariance matrixes of our chosen prior.}

\item{vK}{The virtual counts corresponding to mu of chosen prior}
}
\value{
A \verb{p x p x K} array of matrices sampled from the inverse-wishart distribution where the kth matrix
denotes the posterior covariance matrix of the kth cluster.
}
\description{
Samples an array of matrices denoting the covariance matrices for each cluster
}
\details{
The array of matrices that is returned is  \verb{Sigma_k | A \\ \\\{theta_k\\\}} which had been sampled from\preformatted{`Inverse-Wishart(\\sum\{\\Theta_i - mu_k\},N + vK)` where N is a vector containing cluster counts
, p is a vector of the sizes of theta and 1 is a vector of 1's. Assume that the dimension requirements
are met
}
}
