% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_theta}
\alias{update_theta}
\title{Conditional update of theta using the random effects multivariate normal}
\usage{
update_theta(y, V, mu, Sigma, z)
}
\arguments{
\item{y}{A \verb{n x p} matrix. Each row of y is an observation on p dimensions of the data}

\item{V}{A \verb{p x p x n} array of covariance matricies for each observation of y.}

\item{mu}{A \verb{K x p} matrix. The k-th row of mu is the mean of theta in class k.}

\item{Sigma}{A \verb{p x p x K} array of covariance matricies. The k-th matrix is the covariance matrix for theta in class k.}

\item{z}{A lenght \code{n} vector for the class observations of y.}
}
\value{
A \verb{n x p} matrix. A random sample from the random effects multivariate normal distribution. If the return value is called theta then the details are as follows:
\deqn{\theta ~ N(\mu, \Sigma)}
\deqn{y | \theta ~ N(\theta, V)}
}
\description{
Conditional update of theta using the random effects multivariate normal
}
\details{
\code{n} is the number of observations of y. \code{p} is the dimension of each theta, i.e. the number features of the data. \code{K} is the number of possible classes of y.
}
