% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-functions.R
\name{update_z}
\alias{update_z}
\title{Samples group memberships from a multinomial distribution conditioned on the other parameters}
\usage{
update_z(theta, mu, Sigma, rho)
}
\arguments{
\item{theta}{An 'n x p' matrix, where row i is the parameters of observation i.}

\item{mu}{An 'K x p' matrix where the row k is expected value of theta if theta is in class k.}

\item{Sigma}{A'p x p x K' array of variance-covariance matrices for each class k.}

\item{rho}{An 'K x 1' vector of class membership probabilities}
}
\value{
An 'n x 1' vector of updated class memberships, or if give.Lambda=TRUE, returns a list that contains the 'n x 1' updated class memberships and the computed 'n x k' Lambda matrix.
}
\description{
Samples group memberships from a multinomial distribution conditioned on the other parameters
}
